Reorganise the Tool modules
