+++ AFAAS/core/tools/__init__.py
@@ -1,3 +1,7 @@
 from __future__ import annotations
 
 """The command system provides a way to extend the functionality of the AI agent."""
+
+from AFAAS.core.tools.simple import SimpleToolRegistry
+from AFAAS.core.tools.tools import Tool
+
