+++ AFAAS/core/tools/simple.py
@@ -11,14 +11,14 @@ from AFAAS.lib.sdk.logger import AFAASLogger
 LOG = AFAASLogger(name=__name__)
 
 if TYPE_CHECKING:
-    from AFAAS.interfaces.agent import BaseAgent
+    from AFAAS.interfaces.agent.main import BaseAgent
 
-from AFAAS.configs import Configurable, SystemConfiguration
-from AFAAS.core.tools.base import BaseTool, BaseToolsRegistry, ToolConfiguration
-
+from AFAAS.configs.schema import Configurable, SystemConfiguration
 # from AFAAS.core.tools.builtins import BUILTIN_TOOLS
-from AFAAS.core.tools.schema import ToolResult
 from AFAAS.core.tools.tool_decorator import AFAAS_TOOL_IDENTIFIER
+
+from AFAAS.interfaces.tools.base import BaseTool, BaseToolsRegistry, ToolConfiguration
+from AFAAS.interfaces.tools.schema import ToolResult
 from AFAAS.interfaces.adapters import (
     AbstractChatModelProvider,
     CompletionModelFunction,
